daemon/requirements.txt
fsspec==2022.2.0
GitPython==3.1.27
pika==1.2.0
PyMySQL==1.0.2
SQLAlchemy==1.4.32
